made some updates

+++ appointments/src/EditAppointments.java
@@ -9,7 +9,6 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.geometry.HPos;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
@@ -20,8 +19,6 @@ import javafx.scene.image.ImageView;
 import javafx.scene.layout.Background;
 import javafx.scene.layout.BackgroundFill;
 import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.ColumnConstraints;
-import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
@@ -43,10 +40,16 @@ public class EditAppointments {
     private SceneSwitcher switcher;
     private Business businessLoggedin;
     private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd hh:mm:ss a z");
+    Label errorLabel = new Label();
+    private Appointment appointment;
 
-    public EditAppointments(Stage primaryStage, Appointment appointment, List<Appointment> appointmentListForDay){
+    public EditAppointments(Stage primaryStage, Appointment appointment, List<Appointment> appointmentListForDay, Label label){
 
         businessLoggedin = (Business) primaryStage.getUserData();
+
+        errorLabel = label;
+
+        this.appointment = appointment;
 
         switcher = new SceneSwitcher(primaryStage);
     
@@ -57,7 +60,7 @@ public class EditAppointments {
         BorderPane layout = new BorderPane();
         layout.setLeft(sidebar);
 
-        VBox center = mainPage(primaryStage, appointment, appointmentListForDay);
+        VBox center = mainPage(primaryStage, appointmentListForDay);
         layout.setCenter(center);
 
         editAppointmentsPage = new Scene(layout, 600, 500);
@@ -131,7 +134,7 @@ public class EditAppointments {
         return sidebar;
     }
 
-    public VBox mainPage(Stage primaryStage, Appointment appointment, List<Appointment> appointmentListForDay){
+    public VBox mainPage(Stage primaryStage, List<Appointment> appointmentListForDay){
         VBox center = new VBox();
         center.setAlignment(Pos.TOP_CENTER);
         Button submitButton = new Button("Submit");
@@ -142,12 +145,12 @@ public class EditAppointments {
             boxs[i].setAlignment(Pos.CENTER_LEFT);
         }
 
-        Label typeLabel = new Label("\t\t\t\t\tAppointment Type:  ");
+        Label typeLabel = new Label("\t\t Appointment Type: ");
         TextField type = new TextField(appointment.getType());
         String userType = type.getText();
         boxs[0].getChildren().addAll(typeLabel, type);
 
-        Label dateLabel = new Label("\t\t\t\t\tDate (MM/DD/YY) :  ");
+        Label dateLabel = new Label("\t\t Date (MM/DD/YY) : ");
         ZonedDateTime startDate = ZonedDateTime.parse(appointment.getStartDate(), formatter);
         ZonedDateTime endDate = ZonedDateTime.parse(appointment.getEndDate(), formatter);
         String amPM = "AM";
@@ -155,39 +158,40 @@ public class EditAppointments {
         ObservableList<String> months = FXCollections.observableArrayList(
             "01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"
         );
-        final ComboBox month = new ComboBox(months);
+        final ComboBox<String> month = new ComboBox<String>(months);
         month.setPromptText("Month");
         if(startDate.getMonthValue() < 10){
             month.setValue("0" + Integer.toString(startDate.getMonthValue()));
         }else{
-            month.setValue(startDate.getMonthValue());
+            month.setValue(Integer.toString(startDate.getMonthValue()));
         }
 
         ObservableList<String> days = FXCollections.observableArrayList(
             "01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31"
         );
-        final ComboBox day = new ComboBox(days);
+        final ComboBox<String> day = new ComboBox<String>(days);
         day.setPromptText("Day");
         if(startDate.getDayOfMonth() < 10){
         day.setValue("0" + Integer.toString(startDate.getDayOfMonth()));
         }else{
-            day.setValue(startDate.getDayOfMonth());
+            day.setValue(Integer.toString(startDate.getDayOfMonth()));
         }
 
         ObservableList<String> years = FXCollections.observableArrayList(
             "2023", "2024", "2025", "2026", "2027", "2028", "2029", "2030", "2031", "2032", "2033", "2034"
         );
-        final ComboBox year = new ComboBox(years);
+        final ComboBox<String> year = new ComboBox<String>(years);
         year.setPromptText("Year");
-        year.setValue(startDate.getYear());
+        year.setValue(Integer.toString(startDate.getYear()));
 
         boxs[1].getChildren().addAll(dateLabel, month, day, year);
 
-        Label timeLabel = new Label("End Time: (HH:SS)");
+
+        Label timeLabel = new Label("\t\t  Start Time (HH:SS): ");
         ObservableList<String> hours = FXCollections.observableArrayList(
             "01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"
         );
-        final ComboBox hour = new ComboBox(hours);
+        final ComboBox<String> hour = new ComboBox<String>(hours);
         hour.setPromptText("Hour");
         int tempHour = startDate.getHour();
         if(tempHour > 12){
@@ -195,116 +199,132 @@ public class EditAppointments {
             if(tempHour < 10){
                 hour.setValue("0" + Integer.toString(tempHour - 11));
             }else{
-                hour.setValue(tempHour - 11);
+                hour.setValue(Integer.toString(tempHour - 11));
             }
 
         }else{
             if(tempHour < 10){
-                hour.setValue("0" + Integer.toString(tempHour + 1));
+                if(tempHour == 0){
+                    hour.setValue("12");
+                }else{
+                    hour.setValue("0" + Integer.toString(tempHour));
+                }
             }else{
-                hour.setValue(tempHour + 1);
+                hour.setValue(Integer.toString(tempHour));
             }
         }
 
         ObservableList<String> mins = FXCollections.observableArrayList(
             "00", "15", "30", "45"
         );
-        final ComboBox min = new ComboBox(mins);
+        final ComboBox<String> min = new ComboBox<String>(mins);
         min.setPromptText("Min");
         if(startDate.getMinute() < 10){
             min.setValue("0" + Integer.toString(startDate.getMinute()));
         }else{
-            min.setValue(startDate.getMinute());
+            min.setValue(Integer.toString(startDate.getMinute()));
         }
 
         ObservableList<String> time = FXCollections.observableArrayList(
             "AM", "PM"
         );
-        final ComboBox comboBox6 = new ComboBox(time);
+        final ComboBox<String> comboBox6 = new ComboBox<String>(time);
         comboBox6.setPromptText("AM/PM");
         comboBox6.setValue(amPM);
 
         boxs[2].getChildren().addAll(timeLabel, hour, min, comboBox6);
 
-        Label availabilityLabel = new Label("\t\t\t\t\tAvailability:  ");
+        Label availabilityLabel = new Label("\t\t\t      Availability:  ");
          ObservableList<String> availability = FXCollections.observableArrayList(
             "true", "false"
         );
-        final ComboBox comboBoxAvailability = new ComboBox(availability);
+        final ComboBox<String> comboBoxAvailability = new ComboBox<String>(availability);
         comboBoxAvailability.setValue(Boolean.toString(appointment.getAvailability()));
         Boolean userAvailability = Boolean.parseBoolean((String) comboBoxAvailability.getSelectionModel().getSelectedItem());
         boxs[3].getChildren().addAll(availabilityLabel, comboBoxAvailability);
 
-        Label endTimeLabel = new Label("End Time: (HH:SS)");
+        String endAMPM = "AM";
+        Label endTimeLabel = new Label("\t\t    End Time (HH:SS): ");
         ObservableList<String> endHours = FXCollections.observableArrayList(
             "01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"
         );
-        final ComboBox comboBox7 = new ComboBox(endHours);
+        final ComboBox<String> comboBox7 = new ComboBox<String>(endHours);
         comboBox7.setPromptText("Hour");
-        tempHour = startDate.getHour();
+        tempHour = endDate.getHour();
         if(tempHour > 12){
             amPM = "PM";
             if(tempHour < 10){
-                hour.setValue("0" + Integer.toString(tempHour - 11));
+                comboBox7.setValue("0" + Integer.toString(tempHour - 11));
             }else{
-                hour.setValue(tempHour - 11);
+                comboBox7.setValue(Integer.toString(tempHour - 11));
             }
 
         }else{
             if(tempHour < 10){
-                hour.setValue("0" + Integer.toString(tempHour + 1));
+                if(tempHour == 0){
+                    comboBox7.setValue("12");
+                }else{
+                    comboBox7.setValue("0" + Integer.toString(tempHour));
+                }
             }else{
-                hour.setValue(tempHour + 1);
+                comboBox7.setValue(Integer.toString(tempHour));
             }
         }
 
         ObservableList<String> endMins = FXCollections.observableArrayList(
             "00", "15", "30", "45"
         );
-        final ComboBox comboBox8 = new ComboBox(endMins);
+        final ComboBox<String> comboBox8 = new ComboBox<String>(endMins);
         comboBox8.setPromptText("Min");
         if(startDate.getMinute() < 10){
-            min.setValue("0" + Integer.toString(startDate.getMinute()));
+            comboBox8.setValue("0" + Integer.toString(startDate.getMinute()));
         }else{
-            min.setValue(startDate.getMinute());
+            comboBox8.setValue(Integer.toString(startDate.getMinute()));
         }
 
         ObservableList<String> endTime = FXCollections.observableArrayList(
             "AM", "PM"
         );
-        final ComboBox comboBox9 = new ComboBox(endTime);
+        final ComboBox<String> comboBox9 = new ComboBox<String>(endTime);
         comboBox9.setPromptText("AM/PM");
+        comboBox9.setValue(endAMPM);
+
 
         boxs[4].getChildren().addAll(endTimeLabel, comboBox7, comboBox8, comboBox9);
 
-        Label customerLabel = new Label("\t\t\t\t\tCustomer Name:  ");
+        Label customerLabel = new Label("\t\t     Customer Name:  ");
         Label customer = new Label(appointment.getCustomer().getName());
-        boxs[4].getChildren().addAll(customerLabel, customer);
+        boxs[5].getChildren().addAll(customerLabel, customer);
 
         Label costLabel = new Label("\t\t\t\t\tCost:  ");
         TextField cost = new TextField(appointment.getCost());
         String userCost = cost.getText();
-        boxs[5].getChildren().addAll(costLabel, cost);
+        boxs[6].getChildren().addAll(costLabel, cost);
 
+        Appointment tempAppointment = appointment;
         submitButton.setOnAction(e->{
-            Label errorLabel = new Label();
-            if(userType != null && startDate.isAfter(ZonedDateTime.now()) && endDate.isAfter(startDate)){
-                appointment.setType(userType);
-                appointment.setStartDate((String) year.getSelectionModel().getSelectedItem() + "-" + (String) month.getSelectionModel().getSelectedItem() + "-" + (String) day.getSelectionModel().getSelectedItem() + " " + (String) hour.getSelectionModel().getSelectedItem() + ":" + (String) min.getSelectionModel().getSelectedItem() + ":00 " + (String) comboBox6.getSelectionModel().getSelectedItem() + " -05:00");
-                appointment.setAvailability(userAvailability);
-                appointment.setCost(userCost);
-                changeAppointment(primaryStage, appointment, appointmentListForDay);
+            tempAppointment.setStartDate((String) year.getSelectionModel().getSelectedItem() + "-" + (String) month.getSelectionModel().getSelectedItem() + "-" + (String) day.getSelectionModel().getSelectedItem() + " " + (String) hour.getSelectionModel().getSelectedItem() + ":" + (String) min.getSelectionModel().getSelectedItem() + ":00 " + (String) comboBox6.getSelectionModel().getSelectedItem() + " -05:00");
+            tempAppointment.setEndDate((String) year.getSelectionModel().getSelectedItem() + "-" + (String) month.getSelectionModel().getSelectedItem() + "-" + (String) day.getSelectionModel().getSelectedItem() + " " + (String) comboBox7.getSelectionModel().getSelectedItem() + ":" + (String) comboBox8.getSelectionModel().getSelectedItem() + ":00 " + (String) comboBox9.getSelectionModel().getSelectedItem() + " -05:00");
+            //System.out.println(tempAppointment.getStartDate());
+            //System.out.println(tempAppointment.getEndDate());
+            if(userType != null && tempAppointment.stringToStartDate().isAfter(ZonedDateTime.now()) && tempAppointment.stringToEndDate().isAfter(tempAppointment.stringToStartDate())){
+                System.out.println("True");
+                tempAppointment.setType(userType);
+                tempAppointment.setAvailability(userAvailability);
+                tempAppointment.setCost(userCost);
+                changeAppointment(primaryStage, tempAppointment, appointmentListForDay);
             }else if(userType == null){
                 errorLabel.setText("Please enter an appointment name");
-            }else if(startDate.isBefore(ZonedDateTime.now())){
+            }else if(tempAppointment.stringToStartDate().isBefore(ZonedDateTime.now())){
                 errorLabel.setText("Please enter a valid start date");
-            }else if(endDate.isAfter(startDate)){
+            }else if(tempAppointment.stringToEndDate().isBefore(tempAppointment.stringToStartDate())){
                 errorLabel.setText("Please enter a valid end date");
             }
+            System.out.println("Nothing");
         });
 
         deleteButton.setOnAction(e->{
-            delete(appointment, primaryStage, appointmentListForDay);
+            delete(primaryStage, appointmentListForDay);
         });
 
         Label title = new Label("Edit Appointment");
@@ -320,13 +340,14 @@ public class EditAppointments {
         buttons.setSpacing(2);
 
         center.setSpacing(3);
-        center.getChildren().addAll(spacingBuffer1, title, spacingBuffer2, boxs[0], boxs[1], boxs[2], boxs[3], boxs[4], boxs[5], buttons);
+        center.getChildren().addAll(spacingBuffer1, title, spacingBuffer2, boxs[0], boxs[1], boxs[2], boxs[4], boxs[3], boxs[5], boxs[6], buttons, errorLabel);
         return center;
 
     }
 
-    public void changeAppointment(Stage primaryStage, Appointment appointment, List<Appointment> appointmentListForDay){
+    public void changeAppointment(Stage primaryStage, Appointment passedAppointment, List<Appointment> appointmentListForDay){
         List<Appointment> totalAppointmentList = new ArrayList<>();
+        Boolean alreadyExists = false;
         try{
             //set up FileReader
             FileReader fileReaderAccount = new FileReader("appointmentList.csv");
@@ -334,15 +355,21 @@ public class EditAppointments {
             String line = "";
             String[] tempArr;
             Appointment tempAppointment;
+            ZonedDateTime tempEndDate, tempStartDate;
    
             //read in data and determine if appointment already exists
             while((line = br.readLine()) != null){
                 tempArr = line.split(",");
                 tempAppointment = new Appointment(tempArr[0], tempArr[1], tempArr[2], Boolean.parseBoolean(tempArr[3]), Integer.parseInt(tempArr[4]), Integer.parseInt(tempArr[5]), tempArr[6], Integer.parseInt(tempArr[7]));
-   
+                tempEndDate = tempAppointment.stringToEndDate();
+                tempStartDate = tempAppointment.stringToStartDate();
+
                 //keep note if email is found
-                if(appointment.getID() == tempAppointment.getID()){
-                    tempAppointment = appointment;
+                if(tempAppointment.getProvider().getID() == businessLoggedin.getID() && ((passedAppointment.stringToStartDate().isBefore(tempEndDate) && passedAppointment.stringToStartDate().isAfter(tempStartDate)) || (passedAppointment.stringToEndDate().isBefore(tempEndDate) && passedAppointment.stringToEndDate().isAfter(tempStartDate))) && tempAppointment.getID() != passedAppointment.getID() && !(passedAppointment.stringToStartDate().equals(tempStartDate) && passedAppointment.stringToEndDate().equals(tempEndDate))){
+                    alreadyExists = true;
+                }
+                if(passedAppointment.getID() == tempAppointment.getID()){
+                    tempAppointment = passedAppointment;
                 }
 
                 totalAppointmentList.add(tempAppointment);
@@ -353,33 +380,38 @@ public class EditAppointments {
             System.out.println(except);
         }
 
-        try{              
-            FileWriter fileWriterUser = new FileWriter("appointmentList.csv", false);
+        if(!alreadyExists){
+            try{              
+                FileWriter fileWriterUser = new FileWriter("appointmentList.csv", false);
 
-            for(Appointment a : totalAppointmentList){
-                fileWriterUser.write(a.getType() + "," + a.getStartDate() + "," + a.getEndDate() + "," + a.getAvailability() + "," + a.getProvider().getID() + "," + a.getCustomer().getID() + "," + a.getCost() + "," + a.getID() + "\n");
+                for(Appointment a : totalAppointmentList){
+                    fileWriterUser.write(a.getType() + "," + a.getStartDate() + "," + a.getEndDate() + "," + a.getAvailability() + "," + a.getProvider().getID() + "," + a.getCustomer().getID() + "," + a.getCost() + "," + a.getID() + "\n");
+                }
+
+                fileWriterUser.close();
+
+            }catch(IOException except){
+                System.out.println(except);
             }
 
-            fileWriterUser.close();
-
-        }catch(IOException except){
-            System.out.println(except);
-        }
-
-        int index = 0;
-        for(Appointment a : appointmentListForDay){
-            if(a.getID() == appointment.getID()){
-                break;
+            int index = 0;
+            for(Appointment a : appointmentListForDay){
+                if(a.getID() == passedAppointment.getID()){
+                    break;
+                }
+                index++;
             }
-            index++;
+
+            appointmentListForDay.set(index, passedAppointment);
+            switcher.switchToAppointmentsPageBusiness(editAppointmentsPage.getWindow(), primaryStage, appointmentListForDay);
+
+        }else{
+            errorLabel.setText("An appointment already exists at this time");
+            switcher.switchToEditAppointmentsPage(editAppointmentsPage.getWindow(), primaryStage, appointment, totalAppointmentList, errorLabel);
         }
-
-        appointmentListForDay.set(index, appointment);
-
-        switcher.switchToAppointmentsPageBusiness(editAppointmentsPage.getWindow(), primaryStage, appointmentListForDay);
     }
 
-    public void delete(Appointment appointment, Stage primaryStage, List<Appointment> appointmentListForDay){
+    public void delete(Stage primaryStage, List<Appointment> appointmentListForDay){
         List<Appointment> totalAppointmentList = new ArrayList<>();
         int index = 0;
         try{
